added the refreshing flask and its consume action
if the actor picksup a already consumed refreshing flask, the actor cannot consume it.
When dropping the refreshing flask it is checked whether it is already consumed or not.
When a unused refreshing flask is found the consumerefreshingflask action is triggered where the stamina points are increased from 20% of the maxstamina and the item status changes to consumed